fix(fdc): Refactor Dataconnect types into  and update query error code
